Only log top-level effects
